test: add mutation testing for `evaluator`
The evaluator module internally uses all the other modules and hence if
we generate tests for this one then all other modules are automatically
tested.

The features cannot be tested individually because of how the code is
structured and its interdependence between all other user stories and
all of the other code.

The commands used to generate tests is:
```
PYNGUIN_DANGER_AWARE="1" PYTHONHASHSEED="0" pynguin \
    --project-path . --module-name src.evaluator    \
    --output-path 259045253/test/whitebox/mutation_testing      \
    --algorithm RANDOM --assertion-generation MUTATION_ANALYSIS \
    --seed 0 --maximum-search-time 60 -v            \
    --coverage-metrics BRANCH --create-coverage-report True
```

And the command used to generate coverage report is:
```
coverage run -m pytest  \
    259045253/test/whitebox/mutation_testing/test_src_evaluator.py  \
    && coverage report  \
    -m > 259045253/test/whitebox/mutation_testing/coverage.txt
```